requirements.txt
beautifulsoup4
requests
pandas

apache-airflow-providers-mysql
apache-airflow-providers-mongo
apache-airflow-providers-google
apache-airflow-providers-http
apache-airflow-providers-slack
apache-airflow-providers-amazon
apache-airflow-providers-postgres

# Will pin the correct versions  to these providers soon
